changed function application to fetch the function separately
Before this change, accessing a variable like "display" and "newline"
would insert code to call the corresponding function with whatever was
on top of the stack, and compiling a function application simply
consisted of compiling its arguments followed by the code to fetch the
variable.

This had a couple of problems:
- you couldn't access "display" and the like as variables;
- "display" and the like couldn't tell how many arguments they were
  being given, so they fetched some hard-coded number of arguments
  instead of looking at the source;
- if you did somehow manage to get a pointer to a procedure into a
  variable, then you couldn't actually call it.

So now they do the right thing: compile-ration-2 and
compile-application first compile the code for the arguments, then the
code for the function, then call compile-apply to compile the code to
invoke the function on the correct number of arguments.

+ and - are still open-coded, so they don't need compile-apply to be
called after them; so now they are part of a special syntax list,
along with "begin" and "if".  As an extra special bonus, they now
complain at compile-time if they get the wrong number of arguments,
instead of compiling to code that crashes.

darcs-hash:20080205151231-c0035-eaddeea4c66921acfe3a818188d0e1eceddc77ab