worked over the parser; now it handles EOF in string literals
Each parse routine typically has to dispatch on the next character,
and in many cases, also add that character to a data structure.  Such
nonlinear uses of data require, in Scheme, that we bind the character
to a variable so we can use it more than once.  So there were a bunch
of let-expressions.

The parser from Darius Bacon's "ichbins", which is truly worth
reading, used a technique to avoid the need for let-expressions or
closures: the caller of each routine that needed to see the next
character would simply fetch the next character before calling that
routine.

So I just modified this parser to use a similar technique in places.
As a result, the parser is slightly clearer (I think), and (probably
because my implementation of "let" is expensive) the compiler seems to
run noticeably faster.

darcs-hash:20080222072255-c0035-ec62f25328edf2f92c810a2cd0fe87700609f469